validation: move state into logic module
